Changed the project name from 'ranker' to 'rankr'.
