+++ README.md
@@ -1,3 +1,3 @@
-# Ranker
+# rankr
 
 Scripts for crawling academic ranking websites
