--- ranker/db_models/__init__.py
@@ -1,7 +1,0 @@
-from ranker.db_models.acronym import Acronym  # noqa
-from ranker.db_models.alias import Alias  # noqa
-from ranker.db_models.base import Base, engine, SessionLocal  # noqa
-from ranker.db_models.institution import Institution  # noqa
-from ranker.db_models.link import Link  # noqa
-from ranker.db_models.ranking import Ranking  # noqa
-from ranker.db_models.type import Type  # noqa
