+++ rankr/db_models/__init__.py
@@ -1,0 +1,7 @@
+from rankr.db_models.acronym import Acronym  # noqa
+from rankr.db_models.alias import Alias  # noqa
+from rankr.db_models.base import Base, engine, SessionLocal  # noqa
+from rankr.db_models.institution import Institution  # noqa
+from rankr.db_models.link import Link  # noqa
+from rankr.db_models.ranking import Ranking  # noqa
+from rankr.db_models.type import Type  # noqa
