+++ rankr/db_models/acronym.py
@@ -2,18 +2,21 @@ from sqlalchemy import Column, ForeignKey
 from sqlalchemy.orm import relationship
 from sqlalchemy.types import Integer, String
 
-from ranker.db_models.base import Base
+from rankr.db_models.base import Base
 
 
-class Alias(Base):
-    __tablename__ = "alias"
+class Acronym(Base):
+    __tablename__ = "acronym"
 
     id = Column(Integer, primary_key=True, autoincrement=True)
     institution_id = Column(Integer, ForeignKey("institution.id"))
-    alias = Column(String(255), nullable=False)
+    acronym = Column(String(255), nullable=False)
 
     # Relationships
-    institution = relationship("Institution", back_populates="aliases")
+    institution = relationship("Institution", back_populates="acronyms")
 
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
+
+    def __repr__(self):
+        return self.acronym
