+++ rankr/db_models/link.py
@@ -4,7 +4,7 @@ from sqlalchemy import Column, ForeignKey
 from sqlalchemy.orm import relationship
 from sqlalchemy.types import Enum, Integer, String
 
-from ranker.db_models.base import Base
+from rankr.db_models.base import Base
 
 
 class LinkTypeEnum(enum.Enum):
